// shared/schema/user.d.ts
export interface UserSubscribedCity {
    city: string;
    country: string;
    subscribedAt: string;
}